Three changes:
1) Line 68: add an atomic block so as to replace "for (int i=k+1;i<m;i++) m_rowScore[i]=n-k-1;" by "for (int i=k+1;i<m;i++) atomic m_rowScore[i]=n-k-1;"

2) Line 90: add an atomic block so as to replace "for (point [i] : Rkm) atomic m_rowScore[i]=1;" by "for (point [i] : Rkm) atomic m_rowScore[i]=1;"

3) Line 93: add an async so as to replace "atomic m_rowScore[i]++;" by "async (m_rowScore.distribution[0]) atomic m_rowScore[i]++;"

Of these changes, 3) was the primary source of the bug in v1.7.  Note that m_rowScore[*] is a local array allocated entirely in the place 0.  However the "atomic m_rowScore[i]++;" statement in line 93 occurs in the body of an ateach which means that it can be executed at different places.  In v1.7, this violated the X10 rule that only place-local mutable data can be accessed within an atomic block.  By adding the async, we ensure that all operations on m_rowScore[i] are performed at place 0.


git-svn-id: svn://svn.code.sf.net/p/x10/code@3493 a326200e-df2c-42e0-81fa-597e909af41d