Reimplemented a big chunk of the ast externalization code, catching some missing functionality in the process.  Support reading XML files (.x10ml) from the command-line and from the source path.
Overrode the parser factory method to check if the file is an .x10ml file, and
if so use the XML reader rather than the X10 parser.  Modified scheduler to not
run any AST passes on XML-derived ASTs until after type checking.  Added a
simple verifer pass to check that AST at least has type information.


git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@3946 a326200e-df2c-42e0-81fa-597e909af41d