Working code that runs in parallel. Memory management is in place. All Closures and Frames are appropriately deleted. A debugging compilation flag has been added, which counts the number of constructor and destructor invocation for the objects that are dynamically allocated. This helps with some debugging.
Protocol for deleting frames: A frame created by the user is deleted when the method successfully exits, and the parent has not been stolen. The frames whose parents have been stolen actually end up being the frames in some closure. Such frames are deleted when the closure is destroyed. The remaining frames can be destroyed when popping them and returning from that method.

Note that this requires all frames to be dynamically allocated and delete-able. This would not work when they perform additional functionality and cannot just be deleted (eg., Vijay's Span implementation). But then, there is some relation between this and copying of frames in global quiescence stuff. So we need to get back to this part once global quiescence algorithm is settled.

Global quiscence has been disabled in a variety of ways, and needs to be fixed. Also need to check the sleeping/waking of threads.


git-svn-id: svn://svn.code.sf.net/p/x10/code@3968 a326200e-df2c-42e0-81fa-597e909af41d