Move without modification of all files in x10runtime to x10/runtime.
This is the first step in merging the Thread/Lock implementations
built by Sreedhar into the rest of the C++ runtime system.

The next stage will be to modify the code to match namespace
conventions and to pick up things like Object, Exception, etc.
from src-cpp/x10/lang.  I'm starting work on that now.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@6866 a326200e-df2c-42e0-81fa-597e909af41d