Added demangling for c++ backtraces.
Added x10aux::throwException and modified much of the code (not x10/runtime) to
use it.  The advantage of using this function is it enables backtraces,
provides a single point for replacing exceptions for other mechanisms (e.g. on
the cell) and provides additional typesafety because its parameter is a
ref<Throwable>.

Added String::bytes... Whatever use that is for I don't know.

Killed some commented code from String




git-svn-id: svn://svn.code.sf.net/p/x10/code@6908 a326200e-df2c-42e0-81fa-597e909af41d