XTENLANG-320: ensure 8 byte alignment of rails of doubles on linux
Add __attribute__ ((aligned(8))) to the data field of AnyRail to
force 8 byte alignment.  This is slightly overkill when T is a type
whose natural alignment is less than 8 bytes, but in practice wasting
4 bytes per rail should not be a real space issue.  The argument to
aligned must be a compile time constant, so a few obvious attempts to
test on sizeof(T) didn't work.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@7291 a326200e-df2c-42e0-81fa-597e909af41d