XTENLANG-332: Crash in SerializationDispatched with GC enabled.
Allocate the char* backing buffers used to serialize/deserialize
with the system malloc even when building with gc.  These buffers 
are handed down into pgasrt and either are being retained after 
the call returns or are being freed internally to pgasrt (not clear which).
However, after making this change I was able to run 500 iterations of the
test case that was crashing before without a crash, so whatever the reason
this appears to have corrected at least one problem when GC is enabled...



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@7318 a326200e-df2c-42e0-81fa-597e909af41d