Tweak #defines to make it easy to enable ReferenceLogger without enabling BDWGC. Allows bugs in logging code to be debugged independent of GC.
Fix very silly mistake in sizing the backing arrays used by the 
ReferenceLogger to hold escaped references that was resulting in 
memory corruption because the spine/buffers weren't as big as they
were supposed to be.

I think this probably fixes XTENLANG-343, but that probably isn't so exciting.
Looking at this bug has highlighted the fact that I need to build a more
sophisticated refrenece logger that actually keeps track of whether or not
it has logged a reference so we don't rapidly blow out our logging data 
structures.  This was a known issue with the current prototype, but I'd 
underestimated how rapidly some programs would be serializing references.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk/x10.runtime.17@7893 a326200e-df2c-42e0-81fa-597e909af41d