bump bucket size from 16 to 4k.
A size of 16 was only intended for debugging (to force various code paths to be
taken quickly).  At a bucket size of 16, we can only log a total of 16k 
serializations before we trip an assert that we're out of room.  
With 4k buckets we'll be able to log 4 million serializations before 
giving up.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk/x10.runtime.17@7894 a326200e-df2c-42e0-81fa-597e909af41d