XTENLANG-360  Basic support for -O option to x10c/x10c++
Add support for a -O (-optimize) command line argument to x10c/x10c++ that can be used as a master switch to disable/enable other compiler flags.  Currently -O simple disables place checks and causes the C++ backend's postcompiler to be invoked with -O2 -finline-functions.

Add plumbing to build.xml so that if ant is invoked with -Doptimize=true, then children x10c/x10c++ will be invoked with -O and furthermore than c++ compilation of x10.runtime.17/src-cpp will be done with optimization as well. 

No attempt to simultaneously build optimized/unoptimized versions of XRX and have x10/x10c/x10c++ find the appropriate pre-built version of XRX based on whether or not it was invoked with -O.  Leaving that as a future packaging work item to be addressed in a different issue.

git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@8210 a326200e-df2c-42e0-81fa-597e909af41d