WIP XTENLANG-401 : C++ codesize reduction
Change generated RTT code such that instead of defining a subclass of 
x10aux::RuntimeType for every class and then creating an instance of it, 
we instead just create an instance of x10aux::RuntimeType.

This change reduces the size of libx10lib17.a from 140MB to 85MB on my machine
and reduces executable size for typical small X10 programs by 20-25% (both stripped 
and unstripped sizes).



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@8425 a326200e-df2c-42e0-81fa-597e909af41d