Give up on using templatized itables to enable the direct placement of member function pointers into itables.
The resulting layers of template instantiation was resulting
in internal compiler errors in g++ 4.1 under some combinations
of genric classes and interfaces.

For now, we'll live with the static thunks. Once itables are
actually working and interface dispatch is something we decide
we want to optimize, it might make sense to revisit this decision.



git-svn-id: svn://svn.code.sf.net/p/x10/code/branches@10610 a326200e-df2c-42e0-81fa-597e909af41d