Remove names from asyncs.  This naive implementation was serialising strings with every async, causing a 7x bloat in a simple example i just tried.
If we want to name asyncs, we should do so in a way which does not cause extra
communication.  Since names are static, they are available at every place from
the get-go.  The easiest way to do this is to define toString() on closures.
Closures in the c++ backend already have a toString() function which returns
file:line information.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@10695 a326200e-df2c-42e0-81fa-597e909af41d