Support for dep clauses which refer to qualified this. The support is in two parts.
First, when expanding the depcast templates, ensure that references to specials are printed with fully qualified name e.g. |S.this| since an unqualified |this| is going to be scoped  to the new anonymous class being created to contain the code for the cast | new Object() {...}|. This is done by passing a marker in the context.

Second, ensure that the InnerClassRemover correctly rewrites references to this within the body of depclauses in X10CanonicalTypeNodes. Recall that the InnerClassRemover converts instance class members to static class members, creating an extra field (out$) to contain references to the "outer" object. Hence references such as |S.this| have to be replaced with |this.out$|. This has to be done for occurrences of |this| in DepParameterExpr's as well. 

This is accomplished simply by adding a visitChildren method to X10CanonicalTypeNode_c. Care must be taken to ensure that this method is not used by a TyepChecker, because it will throw spurious errors since the context has not been set up correctly yet. (The TypeChecker should visit the DepParameterExpr only during the leave phase, through the typeCheck method.) 

As a result, the following code now correctly compiles:

class S {
  class Inner {
      class Inner2 {
	  def  m(x:S) {
	      val y:S{self==S.this} = x as S{self==S.this};
	  }
      }
  }
}


git-svn-id: svn://svn.code.sf.net/p/x10/code@10878 a326200e-df2c-42e0-81fa-597e909af41d