Fixed several type-checking bugs reported by Joel (or discovered when fixing these bugs).
For some reason, for loop type-checking was incompletely implemented. Code has been uncommented so that the loop is properly type-checked. This revealed several bugs that have been fixed.

For some reason, type bounds on classes were not making it through.

A distinction is now made between upperBound(t) and upperTypeBound(t). The former provides all the upper bounds for the type t which are sources for inheritance of members. The latter includes upper bounds on the type t that are inferred  upper bounds through variance annotations. These additional upper bounds do not actually supply additional members (methods, fields) -- the fields and methods they identify are already implemented by pre-eisting fields/methods on the class -- but need to be upper bounds because the type system is nominal. 

Fixed typebounds so that variance is respected. If X <: Y and M takes a type parameter with a + annotation, then M[X] <: M[Y].

Fixed X10TypeEnv_c.isSubType so that it uses upperTypeBound.

Had to substantially change the implementation of isSubType. When checking Gamma |- S <: T, the constraints in S need to be added to Gamma. This is necessary because T may be parametrized, and the type parameters may need the constraints for their corresponding type-check to be performed.

So now the following typechecks:

class ForBug {
 abstract class A(v: Int) implements Iterable[Int{self==this.v}] {
   def this(x: Int) {
     property(x);
   }
 }
 def f(a: A{self.v == 41}) {
   for (val x: Int{self == 41} in a) ;
 }
}


The following also typechecks:


public class X {
   class Row {}
   abstract class Mat[+T]{T <: Row} implements Iterable[T] {
       { 
	   for (r:Row in this) ;
       }
       incomplete def m(x:T):Void;
   }

}




git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@10899 a326200e-df2c-42e0-81fa-597e909af41d