XTENLANG-484 WIP: Interactions of itables and generic types
Extend the C++ RuntimeType infrastructure so that it retains 
information about type parameters and thier variance. 
This is required to actually do correct subtype testing of generic
types (which in turn is required to get itable dispatch of generic
interfaces to work in the present of subtyping).

This commit just adds most of the supporting declarations to the code
generator and the handwritten C++ files.  None of this information is
actually used at runtime yet (soon...).



git-svn-id: svn://svn.code.sf.net/p/x10/code@10902 a326200e-df2c-42e0-81fa-597e909af41d