Add a canonical field to the RuntimeType and machinery to initialize it as part of RTT initialization.
For non-generic types, the canonical RTT of an RTT instance is itself.
For generic types, the canonical RTT of an RTT instance is the <void,...>
instantiation of the generic type.

Also some minor cleanups and dead code elimination of RTT infrastructure.

This commit puts all the machinery in place to do correct runtime typetests
for generic types, but does not actually use it.  The next commit will
modify RuntimeType::subtypeOf to actually use the new information when 
determining the subtyping relationship of generic types.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@10932 a326200e-df2c-42e0-81fa-597e909af41d