XTENLANG-481
Codegen fix to ClosureCall so that apply() on Future will
find the abstract (Miranda) method in Future instead of 
attempting to invoke it as an interface method.

Codegen for normal invokes was already correct because it is 
making the interface/non-interface decisions based on the type
of the receiver expression, not the type of the method instance's
container.

Now that the codegen fix is in, we can back out the workaround
in Future.x10 applied in r10895 that added an explicit apply method
to sidestep the problems with Miranda methods.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@10939 a326200e-df2c-42e0-81fa-597e909af41d