XTENLANG-432 and XTENLANG-434:
The c++ codegenerator now inserts explicit calls to x10aux::nullCheck
and x10aux::placeCheck.  implicit nullCheck and placeCheck operations
are removed from ref's -> and * operators.

Although the explicit checks are occurring at a subset of the places
that the implicit checks in the ref operator were occurring, 
there are still some obvious and low-hanging places to suppress 
check generation.  In particular, (a) we should never generate the checks 
on 'this' (b) expressions of value types should not need checks 
(c) the result of a new does not need check (d) ...

However, I also did verify that although at O0, g++ doesn't remove 
redundant checks, at O2 it does understand that a failed nullCheck/placeCheck
won't return and therefore it does successfully eliminate redundant checks.
So, that somewhat lessens the pressure to do that optimization ourself
in x10c/c++ (although we may eventually want to do that if we ever specify
a tighter exception model and the checks are hindering other optimizations).



git-svn-id: svn://svn.code.sf.net/p/x10/code@10959 a326200e-df2c-42e0-81fa-597e909af41d