Fixed Xtenlang-495.
Now typeinference works correctly for each one of the following cases

for (p in 0..9) S  // ok

for ((i) in 0..9) S // ok

for ((i,j) in 0..9) S // type-checking error

for (p:Point(1) in 0..9) S // ok

Needed to introduce the notion of thisVar for constraints and types. Needed to correctly define interfaces for Constrained types T{c} -- see discussion in XTENLANG-495.

Needed to fix some bugs in Desugarer -- the correct context was not being passed in. 

git-svn-id: svn://svn.code.sf.net/p/x10/code@10967 a326200e-df2c-42e0-81fa-597e909af41d