Added a rule for recognizing for((i) in e..f) S and replacing it with for (var i=e; i <= f; i++) S. As a fix for XTENLANG-504.
Needs to be moved out to happen in the lowering phase (currently happens during type-checking.) Needs to be done for foreach and ateach as well. Needs to handle the case for(p(i) in e..f) S. Needs to handle multi-dimensional iteration as well. Needs to handle values flowing through that are of the form e..f (though the improvement in this case is likely to be minor since the region will already have been constructed .. unless one determines through analysis that there is no other use of the region).

git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@11002 a326200e-df2c-42e0-81fa-597e909af41d