(1) Generate hashCode redirection and default hashCode methods for structs     (approx. same hack as toString committed in r11340).
(2) Steps towards getting the rest of the code base to the point where
    x10.runtime.RID could become a struct instead of a Value.  
    (a) Tweaks to c++ @Native annotations for hashCode, toString, and className.

    (b) Changes definition of equals so that Ref and Value are both
        collapsed into a single method that takes an Object 
        (instead of a pair of overloeded methods).  Change was necessary to 
        make progress on getting calls to equals in HashMap to do the 
        right thing when HashMap is used with struct version of RID as the key.

I tested FRASimpleDist, HelloWorldPar, and MontyPi and they all ran ok
with the change in the definition in equals.  It's entirely possible
that some test cases are going to break (but I think those are exactly
the test cases that are going to have to break as we move to the 2.0
object model where Value disappears and Object becomes a class instead
of an interface).

Note: RID being a struct still doesn't work due to additional issues
in equals that need further work.  Continuing along this path...



git-svn-id: svn://svn.code.sf.net/p/x10/code@11357 a326200e-df2c-42e0-81fa-597e909af41d