rename className to typeName to match current struct design.
adjust templates so that x10aux::type_name will 
work for structs (tested by manually adding a typeName()
native method to a struct).

Add temporary definitions of typeName to the built-in primitives
(Boolean, Int, etc). Once the front-end changes are made so that 
typeName is defined on all types, these native methods can be removed
because the general mechanism for structs will work for them too.



git-svn-id: svn://svn.code.sf.net/p/x10/code@11386 a326200e-df2c-42e0-81fa-597e909af41d