Patches and build changes so that bdwgc will build on AIX and pass its own self tests.
Note: still note working on AIX, therefore not enabled on AIX.
Linker problem when linking a x10 application: the symbols
in libgc.a are not being found (GC_malloc, etc) so the link
step is failing.  Most likely a problem with exports when
the library was built, but I haven't figured out exactly the
right incantation yet to get the link to succeed.




git-svn-id: svn://svn.code.sf.net/p/x10/code@11387 a326200e-df2c-42e0-81fa-597e909af41d