Fixes to build.xml to cause bdwgc to be built in 64 bit mode. This then lets us link the resulting .a file into the 64 bit X10 executable produced for AIX/PPC by x10c++.
Was able to successfully compile HelloWorld with GC enabled,
and get an almost immediate segfault when running it.

The issue is the the gcc/standard GNU build process 
really likes to build in 32 bit mode on aix/powerpc and
to get configure/make/install to properly build the 64 bit
library we want, you have to whack it over the head.

Painfully relearning the same lessons as I did in 
Jikes RVM five years ago...



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@11396 a326200e-df2c-42e0-81fa-597e909af41d