XTENLANG-429 WIP:
Change x10.lang.String to be a class with global methods

Eliminate big chunks of code in x10aux/class_cast related to 1.7 
Box semantics that is not needed in 2.0.

Fix a couple of minor bugs in the RTT definitions of Throwable
(Value ==> Ref missed in previous commit) and PlaceLocalHandle 
(structs have no parent).



git-svn-id: svn://svn.code.sf.net/p/x10/code@11561 a326200e-df2c-42e0-81fa-597e909af41d