XTENLANG-429 WIP:  Converting values to classes with global fields/methods.
This commit just does a straight conversion (keeping same semantics 
as previous implementation).  Will open a JIRA to go back and 
re-evaluate this.  It's not at all clear to me that we really want 
all of the I/O classes to be fully global.

Note: also changed a silent serialization/deserializaion problem in
the native Stream classes into an assertion failure. In 1.7, we 
serailize these classes, but didn't transmit the backing FILEPtr
objects.  I think what we really want to do here is to not make the 
methods global, but postponing that reworking to be done under 
separate JIRA item.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@11566 a326200e-df2c-42e0-81fa-597e909af41d