XTENLANG-523 : C++ codegen for structs.
Support for structs implementing interfaces.
For each interface that the struct implements, generate
an ithunk class that provides itables and thunk methods that
redirect to the struct_methods.

Manually tested generated code and was able to successfully
do an interface invocation on a struct method via a slightly
tweaked itable invocation sequence.  A subsequent commit will
make the required tweaks to itable invocation.  Also need fix 
to X10 front-end to preserve type-constraints in the class body
(otherwise, you can never actually get get code where you are doing
an interface invocation on a generic type parameter...it currently is
rejected by the typechecker).



git-svn-id: svn://svn.code.sf.net/p/x10/code@11709 a326200e-df2c-42e0-81fa-597e909af41d