Work around apparent front/middle-end bug that surfaces in HPCC/RA where some piece of the compiler still things that (int)=>Long is compatible with (int)=>Object.
Fortunately (or not...) at the C++ level it appears
that x10::lang::Long is also still considered an instanceof 
x10::lang::Object, so the subType tests in outlinedITableLookup
pass, thus enabling the iTable lookup to succeed.

Will open a JIRA to remind us to look into this further once
we finish scrubbing Ref/Value/Boxing code from the system.
I think there's a resonable chance that this bug will go away
as part of the final cleanup.



git-svn-id: svn://svn.code.sf.net/p/x10/code@11812 a326200e-df2c-42e0-81fa-597e909af41d