-- Make RA work with larger table sizes
During a send operation we check if there are many
outstanding sends (this possibly leads to a memory exhaustion
situation). If there are many outstanding sends, we choose to
block until that send is complete, and do as much work as
possible (discover incoming messages, invoke closures, etc.)

We use a static variable in x10rt_send_msg() to avoid a
recursion problem if the invoked closure sends a message and
discovers more work to do, which send more messages ... and so on.



git-svn-id: svn://svn.code.sf.net/p/x10/code@11886 a326200e-df2c-42e0-81fa-597e909af41d