Fix bug in place_local that would cause memory corruption if the program used more than MAX_FAST_ID (defaults to 255) place local handles.
Remember: when indexing into a hash table, one must mod
the hash code by the number of buckets actually in the table 
before indexing the bucket array!

I checked reference_logger.cc (ie, the other hash table I wrote in x10aux)
and it is correctly doing the mod, so I think the bug is confined to just
this code.  



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@11974 a326200e-df2c-42e0-81fa-597e909af41d