Workaround (hack?) for XTENLANG-676.
Sort the list of XRX files discovered by find to impose a 
deterministic order on the files being passed to x10c++, 
thus increasing the odds of deterministic typechecking results.

When 676 is resolved, we might want to remove the sort 
(or decide to leave it as a hedge against similar future problems).



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@12211 a326200e-df2c-42e0-81fa-597e909af41d