Use printf instead of echo, fixes bug on rlsecomp1 where echo -n hi printed "-n hi".
Tolerate non-existence of libxlpgas_{bgp,sockets,lapi}.a instead of just
libxlpgas_lapi.a, even if platform in theory supports them, because the
CUSTOM_PGAS user may not have built all the pgas backends that are supported on
that platform.  Note that this may cause a silent failure if none of the
archives exist.  The debug messages at the end of the build process should help
in working out what happened if the correct properties files never end up in
x10.dist/etc.  Look for XLPGAS_LAPI_EXISTS and friends.  None of this applies
to people using the web tarballs.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@12282 a326200e-df2c-42e0-81fa-597e909af41d