XTENLANG-524 : Quick fix == on structs in Java codegen.
Simply replicate the reflection based implementation of structural equality
from Value to Struct and add plumbing in Equality to call it.  

I'm going to open a separate feature to improve the efficiency of struct
equality in Java codegen by moving the generation of field-by-field equals
methods out of the C++ codegen and into the front-end of the compiler. 
We really should just be generating ASTs for these methods once in the front-end
and not having the code generators know anything about struct equality.



git-svn-id: svn://svn.code.sf.net/p/x10/code@12289 a326200e-df2c-42e0-81fa-597e909af41d