Removed ValueTypes in most places.
Removed Nat from the type system. It was defined as Int. Duh. Just
changing it to UInt does not work. We have decided to keep Rails indexed by int for ease of code generation for Java. This has a ripple effect -- rank for regions and distributions and points then also stays as Int.  Its best to revisit code written using Nat once Nat is purged from the current code base and then reintroduced as UInt.

Introduced new default constructor for classes and structs. If the class or struct defines properties, and does not define a constructor, then the default constructor takes the properties as arguments and contains a call to super() followed by a call to property(...) with all the arguments passed in. This is sensible in a lot more places.

git-svn-id: svn://svn.code.sf.net/p/x10/code@12437 a326200e-df2c-42e0-81fa-597e909af41d