Closer to a correct implementation of newEquals for C++ codegen. Since the argument to equals is going to be boxed up to Any, the implementation of equals is much closer to that of equals in Java. First do an RTT operation to see if there is a type compatability, then do a downcast, extract the boxed value, and do a ==.
Still some work to do, as annoyingly enough, it is now possible to
have numeric types that are == but not equals. For example:
  with the C++ backend you can now get 
    (3.equals(3 as Float) ===> false
    (3 == (3 as Float) ===> true)

Need to think about how to smoothly integrate user-defined conversions
with == and how to avoid doing lots of pointless boxing up to Any.
Can't be done purely by templates in basic_functions.h because the
boxing operation is injected as part of the codegen for the calls to 
equals....



git-svn-id: svn://svn.code.sf.net/p/x10/code@12511 a326200e-df2c-42e0-81fa-597e909af41d