Fixes XTENLANG-756. -- A one-line mistake in place-checking code that meant that place checks were not performed on  receivers of interface types. This caused a number of place-type bugs to surface in XRX. Fixed them per my understanding -- needs to be reviewed by Olivier. -- Fixed X10TypeEnv_c.typeBounds so that it is aligned with the new type system. The includeObject flag should be ignored. Every type parameter now takes Any as an upper bound.
Fixes a traceback that arose because of a ClassCastException when processing an ambiguous proto types. Proto is handled as a flag. This flag needs to be recorded on the ambiguous type, and then communicated to the resolveed top. Fixed a bug in proto type checking so that proto return types for constructors are now permitted.

Restructured closure code -- moved various AST generators and type generators into x10.util.ClosureSynthesizer. Started to document the surprisingly brittle code for processing closures. Needs to be cleaned up and redesigned.

Fixed home processing in Desugarer to align with recent changes to XRX which broke this piece of code. This now breaks the C++ backend -- apparently home() is not defined yet on structs? (Should be easy to fix.) 

Fixed inscrutable error message that was triggered when processing a non-static typedef.

Fixed closure types toString so that the annoying "<anonymous subtype of ...>" is replaced by "...".




git-svn-id: svn://svn.code.sf.net/p/x10/code@12513 a326200e-df2c-42e0-81fa-597e909af41d