Refactoring of the constraints implementation.
The constraint system in x10.constraint is now made more standard. It exposes publically only those methods that should be used by clients of the constraint solver -- this will make it easier to replace this constraint solver with a more standard congruence-closure solver (necessary for integrating other constraint systems in). The notion of promise is hidden within the constraint solver implementation. Additionally, the special treatment of this and self is removed from this solver.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@12780 a326200e-df2c-42e0-81fa-597e909af41d