XTENLANG-1106 : StackOverflow (SegFault) due to infinite regress during initialization of cyclic RTT's.
Rework RTT initialization protocol in the C++ backend to correctly handle
cycles in the RTT graph (can be introduced by generic type where a type
parameter is also a descendant). The code was already trying to handle this,
but the check for recursive initialization came "too late."  The fix is to
break x10aux::RuntimeType::init into two methods initStageOne and initStageTwo.
initStageOne is responsible for detecting cyclic initialization.  It is
called before _initRTT attempts to get any of the RTT objects for 
ancestor/parameter types.  If initStageOne returns false (not recursive), 
then the other input RTT types for the initialization are acquired by 
calling x10aux::getRTT<TYPE> for them and they are passed as arguments 
to RuntimeType::initStageTwo which finishes the job of initializing the 
RuntimeType object.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@13238 a326200e-df2c-42e0-81fa-597e909af41d