Fixed the test.
The bug was that sum was a field, instead of a method.

The initializer for sum was called before the property was initialized, hence it read the value of the property D.a as 0.

Creating separate JIRA to record that we need to work on defining and implementing initialization order for object fields.

git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@13245 a326200e-df2c-42e0-81fa-597e909af41d