Check that the actual return type of the constructor (as established by the call to property in the body of the constructor) is strong enough to entail the given return type of the constructor.
Original logic had some bugs -- need to make sure fields are being translated to XTerms consistently, got rid of spurious "in super type" context that got pushed in when examining class invariant, realized that the class invariant should be checked to entail invariants asociated with all interfaces implemented by the class. This is not being done currently -- opened Jiras. The work required is slightly non-trivial because one must replace properties specified by the interface in the invariant associated with the interface with the actual properties in the class that implement them.


git-svn-id: svn://svn.code.sf.net/p/x10/code@13288 a326200e-df2c-42e0-81fa-597e909af41d