make the x10rt_lgl_internal_barrier function directly accessible from x10rt_front.h API by adding and x10rt_barrier function.
Required by Java+X10RT work where there are 
multiple initialization phases that needed to 
be ordered by an SMPD-style barrier.

Long term, this will probably be replaced by some
notion of teams and a collective operation on a team,
but until we have that richer API, the simple SPMD barrier
is needed to make forward progress on the Java bindings.



git-svn-id: svn://svn.code.sf.net/p/x10/code@13408 a326200e-df2c-42e0-81fa-597e909af41d