Backout partial attempt to teach the compiler that there are "arrays" that aren't subclasses of x10.lang.Array. Turned out to be more trouble than it was worth.
Change implementation of isX10Array() so that it
returns true for subclasses of x10.lang.Array instead of 
expecting an exact match.  In the new array-lib design, 
we are going to see subclasses of x10.lang.Array showing
up in user-level code and we want various loop optimizations,
etc. to be applied to these subclasses as well.



git-svn-id: svn://svn.code.sf.net/p/x10/code@13572 a326200e-df2c-42e0-81fa-597e909af41d