Final step in splitting local and distributed arrays into separate user-level concepts.
x10.array.LocalRectArray is renamed as x10.array.Array
and it's methods are changed to be local. It no longer
has a distribution, but is a simple mapping of the points
in a Region to values of type T.
x10.lang.Array does not have factory methods, instead user code 
should simply use "new" in the standard way.  
There is a constructor that matches each of the old make 
factory methods, so the transformation should be simple.

x10.array.DistArray becomes the top-level type for
distributed (global) arrays.  It still has factory
methods, since I want to keep open the option of 
having multiple DistArray implementation subclasses if it 
turns out that is useful for efficient implementation of
different styles of distributions. 

Port the code in x10.dist/samples to the new APIs.

Still significant work to do in updating the test suite.
I will continue to work on that over the weekend, but 
wanted to get the API changes in before the end of the day.
Expect many Array/Dist test failures until the test suite 
is updated.

There is also still significant code cleanup, organization,
and x10doc to do to reflect the changes in the array library.
Will continue on that after working through test case changes.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@13624 a326200e-df2c-42e0-81fa-597e909af41d