Cheap and sleazy implementation of stacktrace serialization. Serialize the lowlevel trace computed by backtrace and then hope that the receiving place is running the same exectuable, with any dynamically loaded code loaded at the same virtual addresses. In practice on linux-32 at least this seems to be good enough to be usable.
A more correct solution would be to create the ValRail<String> and 
serialize it.  That would require slightly more work since we'd have
to rework the stack trace printing logic.  The other downside is that
it would significantly increase the cost of serializing an exception
that is just going to be get caught & dropped on the other side (maybe
not a serious concern, but something to consider).



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@13684 a326200e-df2c-42e0-81fa-597e909af41d