More robust version of cross-place stack trace serialization.
Restructure code so that the ValRail[String] version of the 
stack trace is cached in the Throwable object when getStackTrace
is called.  Unless compiled with NO_CHECKS, serialization 
of Throwable internally calls getStackTrace to force creation of 
printable version of the StackTrace.  
We unconditionally serialize/deserialize the ValRail<String> so that the 
proper stack trace will be available anywhere the exception object is
available.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@13686 a326200e-df2c-42e0-81fa-597e909af41d