Fix XTENLANG-1223.
The root problem was that the type of a closure literal is an anonymous subclass of an interface. Function types such as (Int,Int)=>Int correspond to interfaces. Polyglot is happy to now believe that the closure literal (x:Int)=>x could in fact be cast to the interface corresponding to (Int,Int)=> Int since it does not realize that anonymous classes are implicitly final.

Fixed so that the final flag is set explicitly on the closure literal's type, and X10ClassDef_c.flags() is changed so it passes this flag through even for anonymous classes. (Polyglot should prolly be fixed so that it knows anonymous classes are implicitly final.)

git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@13757 a326200e-df2c-42e0-81fa-597e909af41d