Fully specialize the implementation of RectRegion and FullRegion so that they are not implemented via extension of PolyRegion.
Significantly simplifies common case operations and fixes 
several test case failures by avoiding latent bugs in the Poly code.
Performance for RectRegion operations and iterations also noticeably 
improved.

I will shortly do another commit adding an Interface to indicate that 
a region can be asked to construct a PolyRegion representation of its
Points.  RectRegion and FullRegion will both implement this interface.
We will use this interface to remove a bunch of "else" clauses that I
have stubbed out as unimplemented in this commit. We will use conversion
of a region to a PolyRegion to support general interoperation with
user-defined regions as well as more general region operations. 



git-svn-id: svn://svn.code.sf.net/p/x10/code@13810 a326200e-df2c-42e0-81fa-597e909af41d