A simple fix to the dynamic call conversion problem.
Recall that the problem arises when performing a dynamic call conversion on an expression e whose type already contains a place constraint. The place constraint is in some cases represented as self.home = _place3457. (These cases are: the body of a global or static method, or in the body of an async or at whose place arguments do not have a name, or the bodies of ateach clauses.)

The problem is that when trying to generate an AST for this constraint (so code can be generated to check for this condition at runtime), there is nothing that can be done with _place3457. 

But note the following. Let the type of such an expression be T. When checking for e.x, we want to check that e is "here". We were doing that by casting e to T!. But this is overkill. It is sufficient to cast it to B! where B is the base type of T, since we carry identities through casts. i.e. in any case the type of (e as B!) will be B{self==x}!, and this means that all the value information about  x that is known statically will in any case be preserved.

What is gained is that it is now ok for T to have a constraint which names a gensym place (since we will not attempt to generate
an AST for such a constraint). 

This commit now fixes the problem in the following code -- no error is generated:

def m() {
   val x = new Box[Boolean](false);
   at (here.next()) 
      at (here.prev())
        x.t = true;
}

The x.t = true; line will be replaced with (x as Box[Boolean]!).t = true.

git-svn-id: svn://svn.code.sf.net/p/x10/code@13904 a326200e-df2c-42e0-81fa-597e909af41d