Fix for XTENLANG-1263.
When performing dynamic call conversion, ensure that the base type of the expression is a subtype of the formal type. Only then generate a dynamic cast, if necessary. This ensures that the dynamic cast comes into play only when the actual parameter is not a subtype of the formal because of the presence of constraints. 

Thus, "Java" static semantics rules for calls/field selections are preserved. Only the additional checks due to constrained types are deferred to runtime (if dynamic call conversion is not disabled).

git-svn-id: svn://svn.code.sf.net/p/x10/code@13928 a326200e-df2c-42e0-81fa-597e909af41d