Implement an IndexedMemoryChunk, a low-level abstraction of an indexed homogeneous chunk of memory. Implement Array on top of IndexedMemoryChunk.
IndexedMemoryChunk is significantly simpler than 
Rail (no length, no iteration, etc. etc) and provides a minimal 
abstraction of memory for other abstractions to be built on top of.
The API exposed by IndexedMemoryChunk is inherently unsafe 
(no provision for bounds checking). 
The assumption is that checking will be done by the higher level 
abstractions like Array that are built on top of this primitive facility.

Checked Array performance using SeqArray1, 2a, 2b and did not see
any degredation.




git-svn-id: svn://svn.code.sf.net/p/x10/code@14607 a326200e-df2c-42e0-81fa-597e909af41d