Move static methods of Rail to the <void> specialization.
Suggestion from Igor to reduce the code space impact of Rail templates.
Didn't seem to make a measurable difference on Linux-x86, but maybe 
it could help xlC and it does bring the hand-written code more in line
with the conventions used by the compiler to invoked static methods 
of generic classes.



git-svn-id: svn://svn.code.sf.net/p/x10/code@14624 a326200e-df2c-42e0-81fa-597e909af41d