WIP on restructuing place/bounds checking code in DistArray.
Eliminating the idiom of a check routine that takes a closure
to evaluate on error and replacing it with an idiom where
a checking routine that returns true/false is called "inline" 
to branch around an error raising routine that is invoked when
the check failed.  The advantage of this is that the error raising
routine can be marked as NoReturn resulting in better code from the 
post-compiler.



git-svn-id: svn://svn.code.sf.net/p/x10/code@14795 a326200e-df2c-42e0-81fa-597e909af41d