WIP on enabling copyTo/copyFrom for x10.array.Array.
Fix bugs in serialization/deserialization of IMC (IndexedMemoryChunk).
The T* is now properly serialized (as just a T* of size_t bytes) with
the IMC.  Obviously, this gives the programmer lots of rope to hang
themselves if they are using an IMC directly.  Calling apply/set on 
an IMC in a remote place will result in attempting to deref a 
foreign T*, leading to mayhem of unpredictable forms...however
we need to serialize the T* so we can do copyTo/copyFrom efficiently.

Improve typeName and toString methods of IMC to include type param and address.

Change raw and rawLength fields of Array to be global.  This has the
effect of carrying the IMC along with the remote reference to the array,
thus enabling copyTo/copyFrom to be defined with minimal magic.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@14891 a326200e-df2c-42e0-81fa-597e909af41d