Fix nasty bug in usage of free/realloc.
In the x10aux namespace, 'free' and 'realloc'
resolve to 'x10aux::free' and 'x10aux::realloc'
which when GC is enabled map to BDWGC free/realloc.

The intent was to call std::free and std::realloc
because the serialization buffers should always be allocated
out of non-GC managed memory.

Also fix likely space leak when compiling with optimization,
the destructor is supposed to call free to free the buffer
even if we have compiled with optimization...



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@14923 a326200e-df2c-42e0-81fa-597e909af41d