Restructure itable initialization sequence so that we no longer have to rely on ensuring that all of the RTTs needed for itable lookups are initialized outside of the itable dispatcher itself.  As a result, can revert r15518 and restore class_cast_unchecked to being a noop.
The key idea is that in an itable for I, we store also
store the address of x10aux::getRTT<I>.  We do not make any
attempt to get the linker to initialize the actual RTT pointers 
in the itable_entry struct to valid values.  Instead, in the off
branch code we add logic to call the cached getRTT<I> function to
fill in the RTT pointer. So, the slow case becomes a little slower,
but the fast case (after the very first use of an itable) is unchanged.



git-svn-id: svn://svn.code.sf.net/p/x10/code@15520 a326200e-df2c-42e0-81fa-597e909af41d