convert RectLayout from a class to a struct.
Closes 1/3 of the performance gap between the Rail and Array based
versions of SSCA1.  Still 20% left, which I think will require the
val fields of a struct to be declared const at the C++ level before it
can be closed. g++ is not optimizing away redundant index calculations
on array accesses because the stride is being loaded from a non-const field
of RectLayout. 



git-svn-id: svn://svn.code.sf.net/p/x10/code@15608 a326200e-df2c-42e0-81fa-597e909af41d