XTENLANG-1693: C++ backend generating extra proxies for inherited overloaded methods
The root cause of the bug is that the code to generate the set of inherited
methods was based on building up a HashSet<List<Type>> of method formals.
This relies on equals actually doing the right thing for List<Type>, especially
in the presence of constraints on the formals (which in X10 2.x are not relevant
for overloading vs. inheritance decisions). Unfortunately, equals doesn't do
the right thing.  I think attempting to fix equals on Type to match the overloading
semantics is not a good idea, so instead I changed the algorithm to use an ArrayList
and to call isSameMethod instead of doing the type comparision of formals locally.
calling isSameMethod should be more robust against any future change in X10 overloading
semantics.



git-svn-id: svn://svn.code.sf.net/p/x10/code@15632 a326200e-df2c-42e0-81fa-597e909af41d