close a potential race condition in itable lookup code.
It was possible for 2 threads to be doing the same itable
lookup concurrently, and for one thread to have initialized
the itable.id field after the "first" time the other thread 
had looked at it in the fast sequence in itables.h but before
the "second" time the other thread looks at it in the slower
double-check loop.  Rather than attempting to ensure that this
race doesn't happen, declare the id field volatile and also
do the re-check in the slow loop unconditionally.

Makes the slow loop a tiny bit slower, but we're already so far
off the performance hot-path when we are executing in this code
that it should not matter (if it does, we have other problems...)



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@15708 a326200e-df2c-42e0-81fa-597e909af41d