Fix the place checking in the type system so that the X10 2.0 machinery for place-checking can be used to track symbolically the place where an object is created, even though there is no run-time representation of this place (i.e. Object no longer has a home property).
(This essentially meant undoing a lot of the code removal in the type checker I did a few days ago in this branch. The key is figuring out precisely what  delta to make to the 2.0 type checker so that it can deal with 2.1 semantics.)

This enables us to infer in many situations that a GlobalRef''s home is in the current place, and hence it can be dereferenced, e.g. in the canonical situation:

class C {
   val root = GlobalRef[C](this);
}

val x = new C().root;
x();

Also disable the "this checker" from complaining if the computed return type of a constructor contains a reference to this.



git-svn-id: svn://svn.code.sf.net/p/x10/code/branches/localClasses@16127 a326200e-df2c-42e0-81fa-597e909af41d