cachedRail optimization take 3.
Restore correct semantics, but give up for no on trying to 
get the constrained type system to support what I need to say.
Tried a bunch of variations, and none of them were supported.
Will explore further at a later time when we are trying to 
actually exploit constraints for optimization.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@16491 a326200e-df2c-42e0-81fa-597e909af41d