Finished implementation of captured lvals for ats. AtAndBack, AtAndBack2, AtAndBackWithObject and AtAndBackWithStruct all pass with the C++ backend.
Key change was to introduce separate classes to hold captured
struct and captured ref lvals.  The issue is that with refs,
a ref<C> really contains a C*, not a C.  So a captured_lval<C> needs
to contain a C**, but a captured_lval<S> where S is a struct must
contain an S*.  Attempting to jam both a T* and a T** into the same
container class was just not a good idea, so now we have distinct
containers for the two different cases.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@16886 a326200e-df2c-42e0-81fa-597e909af41d