remove the 'uncounted' boolean from various copyTo/copyFrom APIs
I didn't understand what was actually required when I did this part
of the API.  What we actually need is a ()=>void function to be 
registered as a notifier if we are going to do an uncounted
copyTo.  Will add that functionality in the right way in 2.1.1.
But the uncounted copyTo without a notifier is not of any 
practical use, so removing it from the API before the 2.1.0 release.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@16975 a326200e-df2c-42e0-81fa-597e909af41d