Three improvements made to the expression flattener:
1) removes parenthesis AST Node around an Expr

2) Does constant folding for Boolean literals.  (This could be done in the early desugarer.)

3) Uses the side effect detector to determine whether an Eval node can be thrown away.

Also, when evaling Expr's that might have side effects, generates a dummy assignent if the Expe's type isn't void.  This was intended to prevent a Java compiler from seeing unreachable code.  It is guarded by a static final boolean XTENLANG_2000.  If this constant is set to false, x10 will not build: the Java compiler generates possibly undefined variable errors.  I am not sure if the problem is with the Java backend or with my recent changes to the constant propagator.

git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@17902 a326200e-df2c-42e0-81fa-597e909af41d