WIP on XTENLANG-1991
Change codegen for interface invocation to invoke a static method
on the interface that encapsulates the itable lookup. This simplifies
codegen, makes a nicer interface for handwritten C++ code that needs to
invoke X10 interfaces, and prepares the way for implementing interfaces
on C++ primitives via template specialization of the interface class.

On the minus side, it removes the possibility of being able to optimize
away redundant itable lookups within the X10 compiler without changing the 
API again (for example, we could take an optional parameter that is an
itable* and only do the lookup when that parameter is NULL).  Future...
Because of this consideration, I intentionally did not change a few pieces
of handwritten code in Rail and IndexedMemoryChunk that were using the old
invocation style, since they had optimized the itable lookup out of a loop.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@18224 a326200e-df2c-42e0-81fa-597e909af41d