XTENLANG-2130 : Disable contant folding of binary operators    on unsigned numeric types.
The polyglot constant folder doesn't understand unsigned operations,
since Java doesn't have them.  Disable constant folding on Binary exprs
if the type of the expr is UnsignedNumberic.

Need to revisit this in future version of X10 when we have a proper
AST hierarchy for constant values instead of storing them as Java values
and then attempting to use Java to implement constant folding for us.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@18959 a326200e-df2c-42e0-81fa-597e909af41d