XTENLANG-2256: recompilation avoidance when .cc is newer than .x10
With the change of r19410 to generate only a single .cc file for each .x10
file, it is again possible to use file stamp comparision of .x10 and .cc files
to determine if the .cc file must be regenerated.

Of course, as soon as we start doing inlining under -O, this all becomes
much more complex again, but it works for now and should continue to 
work in the future for unoptimized files.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@19415 a326200e-df2c-42e0-81fa-597e909af41d