XTENLANG-2281 : properly handle int overflow in RectRegion size calculation.
Size is calculated using longs to hold intermediate values.  
If the final size cannot be represented in an int, then this is encoded
in the sized field and asking for the size of the Region will result
in an UnboundedRegionException (similar to FullRegion).

isEmpty will return false for such a region (since it is not empty).

Somewhat of a stopgap fix.  

When we expand regions to be indexed by longs, we will have to fix this in
a different way.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@19540 a326200e-df2c-42e0-81fa-597e909af41d