More performance improvements in the constraint-solver and type-checker.
Logical variables are now represented with a smaller amount of storage (an object with an int field, wherever possible). A ConstraintMaker is used to delay projecting the store onto the current set of variables until some information might be needed from them. 

typeEquals has been rewritten to check constraints at the end.

git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@19738 a326200e-df2c-42e0-81fa-597e909af41d