XTENLANG-2530 :
The explodePoint rewriting of the ForLoopOptimizer that attempts to explode
points in calls to Array.apply is not updating closure environment information
to reflect the additional captured variables.  This results in incorrect 
code from the C++ backend if a rewritten call is within an async. 
Disabling this particular rewrite until the code can be fixed.
  


git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@20626 a326200e-df2c-42e0-81fa-597e909af41d