The type-checker checks the legality of constraints.  (XTENLANG-1923)
The only permitted constraints are (top level) conjunctons of
constraint atoms.

The only constraint atoms are 

  t1==t2 
  t1!=t2 
  invocations p(t1,..,tn) or t.p(t1,...,tn) of property methods.  

Property invocations are macro-expanded by the type-checker.

A term t can be built up from local variables, from "self" and "this",
and from field selections and literals. (Specifically, terms such as 
!t, t&&t or t||t or t+t or t*t or -t are not permitted.)

The variable named must be immutable.

If the term is of the form self.p, then p must be a property.

A term may not contain a cast. Therefore the programmer must ensure
that the two terms t1 and t2 in t1==t2 and t1 !=t2 are of the same
type. Specifically, if l is of type Long, then l==213L is accepted,
but l==213 will generate an error.


During the computation of whether a method mi overrides a method mj,
the type-checker corectly handles guards for mi and mj (the guard for
mj must  entail the guard for mi). (XTENLANG-2325)

Adjusted BuiltInTypeRules so that !=null constraints are propagated
for IntRange, Region*Region and Range*Range, and we give up on an
attempt to permit t1&&t2 as a constraint term.

A major change is made to the way method instantiations are
computed. The method/constructor instance returned by
Matcher.instantiate now has a fresh local variable for each argument.
The types of arguments in the instance may reference these local
variables and should no longer contain "unbound" uqv's. This should
fix sightings of uqv's in error messages. 

The type-checker now correctly handles implicit conversions in method
calls in case the argument to be converted is not a variable, e.g. 
new Array[Int](1..16, (Point(1))=> 1) now works. The mi passed into
Converter.tryImplicitConversions corresponds to the def this(r:Region,
init:(Point(r.rank))=>T) constructor and now has a fresh local
variable (say arg21) of type IntRange (representing the actual argument 1..16),
and another one (say arg22) representing (Point(1))=>1. tryImplicitConversions
coerces the first to Region(1) (and some other constraints) and then
correctly updates the type of arg21 to Region(1) before attempting to
find a conversion of arg22 to (Point(arg21.rank))=>Int.


git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@20892 a326200e-df2c-42e0-81fa-597e909af41d