An improvement for XTENLANG-2650: memoize results of method check
During StaticInitializer phase the compiler checks the initializing
expression to verify whether that field can be initialized independently
in each place (i.e., the value is deterministic), or whether it needs
to be computed at place 0 and later passed on to other places.

This check includes recursive visiting of the AST for the constructors
and called methods, which can be quite slow for some cases, in part
because some basic constructors are visited many times.

In this patch we introduce cache of the results for two most frequently
used visitors: checkProcedureBody() and getProcedureDeclaration()

Compilation time of HeatTransver_v0.x10 is improved by half,
but still is quite long.


git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@21246 a326200e-df2c-42e0-81fa-597e909af41d