Patch from Josh Milthorpe (ANU) to extend the GlobalRef API with helper functions that encapsulate a useful idiom:
  if (here == ref.home) {
      return eval(ref());
  } else { 
      return at (ref.home) eval(ref());
  }

There's a value in providing this API as an intrinsic method
of GlobalRef because doing so is the only way to avoid a redundant 
dynamic check of the place constraint in the then branch 
(constrained types in X10 2.2 are not path-sensitive).

Also fix a typo in the header guard in GlobalRef.h that prevented 
Josh's original patch from working cleanly.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@21254 a326200e-df2c-42e0-81fa-597e909af41d