WIP on pulling the cast injection out of MessagePassingCodeGenerator and into a separate pass.
This little bit is sufficient to fix the CollectingFinish test case 
regressions, because by putting the casts into the AST before codegen,
the visit to discover the types being used by the AST actually sees the
types in the casts, so we include the necessary header files.

Will finish the job and pull the rest of the unchecked casts being injected
by the codegen into this pass in a subsequent commit either tonight or
Monday.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@21520 a326200e-df2c-42e0-81fa-597e909af41d