XTENLANG-2693 : Segfault on BG/P in Region.makeBanded
Tracked the problem down to C++ level memory corruption following 
shortly after an attempt to deallocate a size 0 indexed memory chunk
which had been allocated as 16 bytes pad + 0 bytes real storage on
an 8 byte boundary and then rounded up to a 16 byte boundary. The dealloc
call was being made with the rounded up pointer, not the one that was 
actually returned from the lower-level alloc routine.

What this fix does is to simply not allocate a backing hunk of memory
for a IMC of size 0, since it will never actually be used.  This then
prevents us from ever attempting to allocate a IMC of 16 or fewer bytes,
which pushes the alloc request into the next size class, which is
naturally aligned at a 16 byte boundary so the rounding is a no-op.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@21575 a326200e-df2c-42e0-81fa-597e909af41d