Added a top-level ant target named "download-dependencies", which will download all the external jar files and bdwgc (if GC is not disabled).  A user can now check out the trunk, run "ant download-dependencies", and then zip up the directory and run the build on another machine which does not have internet access.
This target does not take care of PGAS binaries, because there's no easy way to know which platforms/versions would be appropriate to use (if any), and downloading them all is wasteful.  So the user is still responsible to download these manually into the x10.runtime/x10rt/ folder if they are desired.

git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@21619 a326200e-df2c-42e0-81fa-597e909af41d