XTENLANG-2454 : Precisely define semantics of shift count operand  for operator << >> and >>>.
Shift count is uniformly masked (if required) so that only
the low order number of bits that corresponds to the log of the 
size of the value being shifted are considered as being part of the
shift count.

For ints and longs on the Java backend, no explicit mask is needed
because this is how Java defines shift already.  Mask is not needed
in C++ backend for ints and longs on x86 hardware, but is on powerpc.  
So put the mask in the @Native annotations and rely on the post compiler 
to optimize it away when it is not actually needed.

Masks are needed for smaller values (byte, short) to ensure the 
X10 masking semantics. Runtime cost for the mask should be negligble.

No change needed to language spec, as it already specified the masked
semantics for shift.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@21699 a326200e-df2c-42e0-81fa-597e909af41d