cleanup from r22229.
Never try to select the output stream for a closure class declaration
based on the lexical chain of the closure definition.  In the presence
of inlining, this can't be done correctly based solely on the closure
def.  The stream decisions need to be made when we are doing the
codegen for the top-level entities (constructors, fields, methods)
that might contain a closure since it is at that point that we know
the stream where the body of the construct (which is what uses the closure
class declaration) is being generated.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@22232 a326200e-df2c-42e0-81fa-597e909af41d