Added path() to the cannonical string for an X10ProcedureDef.
Refuse to inline procedures with position == COMPILER_GENERATED.  (Compiler generated methods are fine to inline, but there positions must contain enough informaton to disambiguate them from other methods.)

Refactoring.  TODO: rename InlineCostEstimator to DeclPackage and InlineDeclHarvester to DeclPackager.

There will be two constructors for DeclPackage.  One takes a job and a decl and tries to compute the cost of inlining the decl.  The other takes a reason the reason a decl cannot be inlined.


git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@22340 a326200e-df2c-42e0-81fa-597e909af41d