XTENLANG-1560: constant propagation of closure literals
Implement constant propagation of closure literals in a pre-inlining
constant propagation pass.  This enables many inlining opportunities
by getting a closure literal passed as an argument to an inlined
function to the callsites where it is applied.  This in particular
makes a large performance difference for Array constructors and for
Array operations such as map, scan, and reduce. 

There are two minor issues with this optimization, but since the
performance benefit is significant we are going to document them in
JIRA, but not address them for 2.2.1.
First, the closures are propagated blindly, so for example in
code like:
  val cls = (i:int)=>i+10;
  foo(cls), bar(cls), baz(cls)
we will actually end up with three closures, not one. This increases
code size and impacts runtime performance.  In practice, I think this
is a fairly uncommon idiom (overall code size of XRX slightly
decreased with closure inlining enabled). 
Second, in code like:
  val cls = (i:int)=i+10;
  val cls2 = cls;
  if (cls == cls2) { ... }
constant propagation will result in cls != cls2.
This is not the only place in which reference equality for closures
does not work (eg, wrapper closures injected for up/down casts), so I
don't see it as blocking concern.

Also some additional minor changes related to implementing this feature:
  (a) StmtExpressions are never constants!  Stmts have side-effects!!
  (b) Add @Inline pragmas to Array map, reduce, scan
  (c) remove @Header pragmas from Array.  Not needed for a generic
      type like Array since all methods go in the header stream.
  (d) Add NoInlines to a proplematic Array constructor in
      Point/Region. Will open JIRA to document the problem this causes
      for Managed X10 codegen when attempting to inline the constructors.




git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@22838 a326200e-df2c-42e0-81fa-597e909af41d