Don't try to generate a #include for a Java type.
If the program actually needed the #include (ie, really used the type)
then it will still fail at post-compilation time.
However, if the type was only actually used in a type constraint, 
then it can still be compiled successfully by the postcompiler 
if we just ignore the use of the Java type.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@22977 a326200e-df2c-42e0-81fa-597e909af41d