Even more fixes for the changes to static initialization of perProcess fields.  I think this time it is actually correct.
In the end, we can't actually do that much optimization of perProcess
fields because C++ static initialization happens in an arbitrary order.
Although it is tempting to try to avoid the check of an initialized
boolean when accessing a static field, it is not sound to do so because
the X10 code may be running as the result of static initialization of 
another field and therefore the check is required.  Since we're in 
a purely static compilation world, we have to pay the price for this 
check all the time (vs. in a JVM where it can be removed once the field
has been initialized).  Sigh.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@23657 a326200e-df2c-42e0-81fa-597e909af41d