WIP on eliminating eagerly assigned serialization ids.
Remove $_get_serialization_id from X10JavaSerializable interface;
change compiler to not generate the accessor method in compiled code.

Internally we are still relying on code run as a side-effect of
calling the DeserializationDispatcher on a couple of code paths, so 
can't eliminate the act of assigning the id quite yet.  
Will work on eliminating the few hold out uses next.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@24866 a326200e-df2c-42e0-81fa-597e909af41d