XTENLANG-3245 : Compiling with -DEBUG generates invalid code.
Disable the generation of linemaps, etc. for Native X10.  Both the 
linemap information and the debugger itself need significant work to
support object model changes made in X10 2.3 and X10 2.4.  Since we do not
have plans/resources to do this work, but may want to do it at some future
point we will disable the code, but leave it in place.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@25790 a326200e-df2c-42e0-81fa-597e909af41d