Minimal fix for problem building Native X10 with FLATTEN_EXPRESSIONS.
The issue is that in an @NativeRep class all methods are native; the lack of 
an @Native annotation for c++ doesn't actually cause the compiler to look
at the X10 body of the method/field.

A follow on commit will clean up CheckedThrowable (and other @NativeRep C++
classes) to make all methods/fields native and to remove redundant @Native
annotations (don't need an @Native if it is simply repeating what the compiler
would generate anyways).



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@26401 a326200e-df2c-42e0-81fa-597e909af41d