Remove obsolete StaticFieldAnalyzer and related remnants.
This was an optimization that marked fields with trivial initializers
as per-process.  With the change in X10 2.2.3 to per-process statics,
this optimization became useless.  I'm removing it now, because it actually
is also buggy, and a more precise treatment of cast operations in Native X10
exposed the bug.  The pass was improperly writing:
 static val x:Any = (Any)"abc";
into
 static val x:Any = "abc";
where the type of the X10StringLiteral_c was set to Any{self=="abc"}
instead of String{self=="abc"}.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@26428 a326200e-df2c-42e0-81fa-597e909af41d