XTENLANG-3466: NativeX10 post-compiler error due to spurious code injected during lowering of at.
For certain obscure cases (closure in a generic method that contains no use of
any of the type parameters and therefore does not need to be generic), the 
extra try/catch/hidden-rethrow code injected around the body of an at 
stmt/expr results in invalid generated C++ code. The simplest fix is
to just avoid generating this try/catch/rethrow wrapper in the first place
for NativeX10 since C++ doesn't have baked-in rules about CheckedExceptions
and throws declarations and therefore the wrapper is completely useless 
for Native X10. 



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@28606 a326200e-df2c-42e0-81fa-597e909af41d