SYNTAX CHANGE: method body and constructor body cannot be a block introduce with an equal.
For example, the following method declaration is no more valid:
 def f() = { f(); }
The user must write:
 def f() { f(); }

Method definitions can no more finish with an implicit return.
Declarations of the form
 def f() = {  ... ; expr }
must be rewrite
 def f() { ...; return expr; }

A declaration of the form
  def f() { expr }
can also be rewrite
  def f() = expr

This syntax change is made to improve the syntax error recovery.



git-svn-id: svn://svn.code.sf.net/p/x10/code/branches/parser@29228 a326200e-df2c-42e0-81fa-597e909af41d