Add 'try' operator in MultipleExceptions.
Provide a user-defined control structure
  "MultipleExceptions.try {...} catch (me: MultipleExceptions) {...}"
such that 'me' contains the MultipleExceptions raised by the body of 
the 'try' but without nesting of MultipleExceptions.

It also provides a user-defined control structure
  "MultipleException.try {...} catch (Rail[E]) {...}"
that catches all the exceptions of type E contained in the body of
the MultipleExceptions raised by the body of the 'try'. Remaining 
exceptions are re-thrown.