Clarify where notifier function is invoked in uncountedCopy documentation
Expand method comment to make explicit that the notifier function
will always be invoked in the place of the destination Rail. If the
destination is a Rail, then that is the same Place as where the
uncountedCopy function is invoked.  If the destination is a GlobalRail,
then that is the home location of the GlobalRail.