Add checking for memory overruns on Team collective operations
Addresses a recent user problem where failure in Team to verify that
all reads/writes to Rail src/dst parameters would be in bounds before
calling into native (MPI/PAMI) collectives resulted in memory corruption
and intermittent crashes.  This change puts in explicit bounds checking
logic on entry to the collective operation to ensure we don't invoke
a native collective with invalid offset and/or element counts.  If an error
is detected, a message is printed along with a stack trace to stdout before
the exception is raised.  This is advisable because in typical collective
usage patterns the exception itself will not be reported before the
program hangs (because the thread raising the exception will not
successfully exit the collective operation, thus very likely causing the
dynamically enclosing finish to not terminate and report the exception).