Fix bug in hosts vs. accelerator identification in PlaceTopology
Fix bug introduced in 7442dbc when as part of resilient/elastic
X10RT work the caching of x10rt_nplaces and x10rt_nhosts was
eliminated. The @Native annotation for PlaceTopology.num_primary_places()
was changed to call x10rt_nplaces() when it should have been calling
x10rt_nhosts().  As a result, all of the CUDA sample programs
were not actually using GPUs, because they rely on the PlaceTopology
functions to determine whether or not GPUs are present.