Enhance CUDA tests to detect when GPUs are unexpectedly unavailable.
Add a check in the test driver for all CUDA samples that checks to be
sure that GPUs are actually available for execution if X10RT_ACCELS
has been set in the testing environment.  This should help us flag
errors when we either break CUDA support entirely or the test system
is misconfigured.