XTENLANG-3540: Allow a SparsePlaceGroup to be constructed from any set of Places
This commit lifts the restriction that the Rail[Place] used to construct
a SparsePlaceGroup be sorted in ascending order by Place.id.  This enables
cheaper recover operations for GML, DistArray, etc. in Resilient/Elastic X10
by avoiding the need to re-shuffle all data at all Places with a higher
id than the failed Place when a new Place is injected into a PlaceGroup
to logically replace a failed Place.

XTENLANG-3540 # resolve