XTENLANG-3541: Resilient X10: 'at' unblocked too soon on place failure
Found a bug in existing implementation of resilient at via
an internal finish where a blocked activity could be released
prematurely due to the death of a Place it had shifted through
via a chain of ats.

This commit adds test cases that expose the problem and
fixes the implementation by adding support to the XRX
internals to track the chain of internal finishes created
to implemented the blocking aspect of 'at'.

Also fixed a cut & paste error in TestNestedFinishMiddleFailSync
where main was instantiating the wrong Test class.