X10_STABLE_POOL_SIZE must be initialized before XRX starts running.
When I introduced X10_STABLE_POOL_SIZE in 596cf04, I failed to
add it to the list of statics that must be forcible initialized
before XRX is considered to be fully booted.  This results in a
race condition that can hang XRX because this flag is checked
in XRX scheduling internals when holding other internal locks.