Default 'atFinishState' to UNCOUNTED.
The atFinishState of an activity is used internally by runAt
to chain together place-shifted activities. For activities
that were not created to be the place-shifted continuation of
an activity by runAt/evalAt, the atFinishState should be set to
UNCOUNTED_FINISH.  This avoids the unnecessary adoption of the
synthetic internal finishes used by runAt/evalAt by "real"
finishes during place failure processing.