Optimization of state updates for dead places in Place0 resilient finish.
Instead of doing O(NUM_DEAD_PLACES) work on every single call to
waitForFinish and notifyActivityTermination to ensure counters for
dead places are zeroed, add guards on counter increments to ensure
that counters that have already been zeroed due to place death are
not incremented (due to late arriving spawn/create messages).

Additional cleanups are possible (future commits), but this change is
enough to reduce the cost of activity termination and waitForFinish
to O(1).