More small improvements to Place0 resilient finish
Refactor to fully separate place death processing from quiescence check.

Started streamlining waitForFinish.  Combined termination notification
with getting the latch to wait on to reduce from 3 communications with Place0
to 2.  Still more to do here; after the latch is released we go back to
Place0 again to get the exceptions.  The exceptions (or at least the lack
of them if there are none ) should be rolled into the unlatch message
from Place0 so that when the thread is released it can just go immediately
in the normal case.