Fix bug in 8d78c4e9 related to liveness of ResilientFinishPlace0 instances
The latency reduction optimization of 8d78c4e9 means that we can no
longer mark instances as Mortal because we create GlobalRefs to
instances that are not the root instance of the finish. In non-root
instances waitForFinish will never be called, so we can't tie the
lifetime of the RFP0 instance to any particular activity.  Therefore
we need to manage lifetime via calls to forget when the localCount
goes to 0 in notifyActivityTermination.