haszero constraint only required on make methods without initializer.
We only have to require haszero on the element type of a DistArray
when the user does not provide an initializer.  For make calls with
an initializer, we can be less strict because all user-accessible elements
of the backing rail will be properly initialized to a non-zero value
by the init value/closure.