Update for change in x10.util.HashMap which now requires {haszero V}
Changeset b53ea120f eliminated the use of Box by x10.util.Map
and related classes by adding a haszero constraint to the value
being stored in the map.  Since the ResilientStoreForDistArray
internally uses a HashMap, it needs to have that constraint as well.
The constraint is trivially satisfied by the only client of the
code (which stores Rail[T] as the value; Rail is a class so haszero is true).