Fixed a deadlock when a place fails while there is data buffered to send to it
The sockets transport has a buffering capability, which will buffer outgoing data
if the remote place is not receiving it as quickly as we are sending it.  This is
enabled by default.  If a place fails while data is buffered, the worker thread
which attempts to flush the buffer will instead get stuck in an infinite loop, as
the data is never removed from the buffer queue, and can't be flushed.  This
change discards any buffered data associated with a dead place.