New attempt to prevent segfaults in team under Resilient X10
The previous commit to Team.x10 attempted to eliminate segfaults
that were caused under resilient X10 when a place failed in the middle
of a collective data transfer.  The fix was imperfect, because it checked
if a team was valid before making the transfer, but the place failure,
and marking of the team as invalid, may occur after the check, but before
the actual data transfer.

This commit reverts the previous one, and provides an alternative
implementation which checks for null Rail fields, before writing to
them or casting them to non-null before creating a global rail from
them.  Not all data transfer points contain checks for null, but all of
them that have come out in testing are covered.  The remainder likely
don't need them, although it's possible that a future segfault will show
that to be incorrect.