Third attempt to fix segfaults in Team during place failure
Local rail objects used in collective operations may be null if a place has failed, but they may also be null if the collective simply doesn't use the argument for all places.  For example, the source in a broadcast is only needed in the root.  So if a rail is null, we don't want to use it in a Rail.asyncCopy or similar operation, but we can't skip the rest of the steps.