Team: don't forget GlobalRails while they're still in use
In Team.LocalTeamState.collective_impl, don't forget GlobalRails until
any asyncCopy operations on them are guaranteed to be completed
(because the Team has moved on to the next phase at the source/target
of the copy). Also don't create dummy Rails as targets for GlobalRails;
this is likely to hide bugs (we should never be attempting a copy
from/to a null/empty Rail).
Reduce lexical scope of GlobalRails for copy from parent to child.