x10.xrx.Pool: avoid race on workers.count when starting new workers
Team collective tests were crashing with x10rt=sockets with a NPE or segfault
in Thread::start().  The problem seems to have been a race condition between
the loop in Pool.operator this(n:Int), which starts the initial threads, and
the newly created threads calling Runtime.increaseParallelism(), which
modifies the value of workers.count.
Changed Pool.operator this(n:Int) to save the value of workers.count before
starting any of the new threads.