Serializer of the parser state.
The options ANTLR_CACHE_WRITE and ANTLR_CACHE_READ can be set to 
save and restore the state of the parser (in a file parser.cache).
By default these options are unset.

This commit introduces a dependency on the Kryo library. Therefore
we have to add in x10.dist/lib the following jar files:
 - kryo-3.03.jar
 - minlog-1.3.0.jar
 - objenesis-2.1.jar
 - reflectasm-1.10.1-shaded.jar