XTENLANG-3546: Segfault in NativeX10 on upcast of function type
Added a missing case to the CastInjector: when an explicit cast
is used to upcast one function type to another in Native X10 we
need to inject a wrapper closure to convert the arguments/return
type to the expected types. In effect, a successful cast expression
needs to be treated like an assignment statement wrt. boxing upcasts
of function types.