determine if using xlC from value of X10LIB_CXX
Previously we assumed the postcompiler was xlC if the compiler name
was any of xlC, mpCC or xlcxx.  However, on newer POWER systems the
MPI compiler is called mpicxx, just like on other non-POWER systems.
To avoid confusion, determine if postcompiler is xlC by checking the
value of X10LIB_CXX.