X10RT MPI: reduce number of lock/unlock cycles
Eliminate a couple of unnecessary queries on the start of the request queue.
Use the TeamPostprocessDB lock to control polling directly, rather than using
it to control a boolean variable that says whether polling is happening.
If the lock is held, polling is underway; otherwise, keep trying the lock
until we manage to lock it.
Also eliminated some unnecessary copying of the req pointer.