Hazelcast: add support for thin places (HC clients)
Basic support for allowing a subset of X10's places to be
hazelcast clients instead of full-fledged members of the
hazelcast grid.  This allows us to run multiple places on
a single node while still being resilient to node failures
(by mapping places such that only one full-fledged
hazelcast place is located on each node).

The initial configuration is simple: Places 0...env(X10_NUM_HAZELCAST_INSTANCES)-1
are full cluster members, Places env(X10_NUM_HAZELCAST_INSTANCES)..X10_NUM_PLACES
are hazelcast clients that connect to the cluster, but are not cluster members.

If X10_NUM_HAZELCAST_INSTANCES is not set, or is set to be > X10_NUM_PLACES,
or is set to be less than 1, then we get the default behavior of every
place being a full cluster member.