bug fix: Elastic X10 (via YARN) and Hazelcast
When a new Place is elastically joined by YARN to an already running
X10 computation it will be told the number of places and its place
number by the launcher.  Thus it will join the X10 computation by
sending HELLO messages to each Place (not by the PLACE_REQUEST/PLACE_RESPONSE
protocol).  Therefore we need to replicate the CONNECT_DATASTORE check
into the handling of HELLO.  Failure to do this when running with
Hazelcast results in elastically adding places not connecting to the
Hazelcast store, which then results in failures/hangs because they cannot
participate in finish state updates.