fixes to support resilient Team operation
Instead of maintaining a separate boolean field Team.LocalState.isValid,
simply treat invalidity as another Team state. Once a Team reaches the invalid
state, it cannot return to a valid state.
If X10_TEAM_DEBUG_INTERNALS is set, log in sleepUntil whenever a thread starts
waiting on a condition that is not already satisfied, and log again when
the condition is satisfied, or if the condition failed because the Team has
since been marked invalid.
Don't ever unconditionally set the phase variable (as this could allow the
Team to return from an invalid to a valid state); instead, always use
compareAndSet(EXPECTED_STATE, NEW_STATE).
Use counted instead of uncounted asyncs during the upward pass from child to
parent; likewise, use Rail.asyncCopy instead of Rail.uncountedCopy.
Don't attempt any copies if the parent place has been marked invalid.
Previously, after failing out of a Team operation, each place would attempt to
notify its immediate parent and children of the failure.  However, if the
failure was due to either the parent or one of the children, it was possible
that the grandparent/grandchildren would not be informed of the failure.
Therefore, if a parent/child is marked as dead, notify the grandparent/
grandchildren accordingly.