alter Team to use one activity per place
A different approach to supporting resilient Team operation.
Previously each place in a Team operation required up to two standard remote
activities (one to copy/notify upwards to parent and another to copy data
down from parent), plus an uncounted activity to notify the child when the
Team operation is complete.
Changed to use a single standard remote activity from child to parent (with
optional async copies), which waits until the parent has completed its portion
of the Team operation before returning to the child.

Other changes:
Instead of maintaining a separate boolean field Team.LocalState.isValid,
simply treat invalidity as another Team state. Once a Team reaches the invalid
state, it cannot return to a valid state.
If X10_TEAM_DEBUG_INTERNALS is set, log in sleepUntil whenever a thread starts
waiting on a condition that is not already satisfied, and log again when
the condition is satisfied, or if the condition failed because the Team has
since been marked invalid.
Don't ever unconditionally set the phase variable (as this could allow the
Team to return from an invalid to a valid state); instead, always use
compareAndSet(EXPECTED_STATE, NEW_STATE).
Previously, after failing out of a Team operation, each place would attempt to
notify its immediate parent and children of the failure.  However, if the
failure was due to either the parent or one of the children, it was possible
that the grandparent/grandchildren would not be informed of the failure.
Therefore, if a parent/child is marked as dead, notify the grandparent/
grandchildren accordingly.